fix: resolve clang-tidy warnings in TP code
- tp_reassembler.cpp: use auto with static_cast (hicpp-use-auto)
- tp_segmenter.cpp: explicit size_t cast to silence implicit widening
  of multiplication result (bugprone-implicit-widening-of-multiplication-result)

+++ src/tp/tp_segmenter.cpp
@@ -107,7 +107,7 @@ TpResult TpSegmenter::create_multi_segments(const Message& message,
 
     // max_segment_size is the payload capacity (excluding headers).
     // Round down to a multiple of 16 for non-last segments.
-    const size_t uniform_payload = (config_.max_segment_size / 16) * 16;
+    const size_t uniform_payload = static_cast<size_t>(config_.max_segment_size / 16) * 16;
     if (uniform_payload == 0) {
         return TpResult::SEGMENTATION_FAILED;
     }
